move and rename things
